+++ core/src/main/java/org/traffichunter/titan/core/channel/ChannelPrimaryIOEventLoopGroup.java
@@ -25,8 +25,8 @@ package org.traffichunter.titan.core.channel;
 
 import lombok.extern.slf4j.Slf4j;
 import org.traffichunter.titan.core.concurrent.*;
-import org.traffichunter.titan.core.util.concurrent.Promise;
-import org.traffichunter.titan.core.util.concurrent.ScheduledPromise;
+import org.traffichunter.titan.core.concurrent.Promise;
+import org.traffichunter.titan.core.concurrent.ScheduledPromise;
 
 import java.util.ArrayList;
 import java.util.List;
