+++ core/src/main/java/org/traffichunter/titan/core/channel/ChannelSecondaryIOEventLoopGroup.java
@@ -27,8 +27,8 @@ import org.traffichunter.titan.core.concurrent.EventLoop;
 import org.traffichunter.titan.core.concurrent.EventLoopFactory;
 import org.traffichunter.titan.core.concurrent.IOEventLoop;
 import org.traffichunter.titan.core.concurrent.IOSelector;
-import org.traffichunter.titan.core.util.concurrent.Promise;
-import org.traffichunter.titan.core.util.concurrent.ScheduledPromise;
+import org.traffichunter.titan.core.concurrent.Promise;
+import org.traffichunter.titan.core.concurrent.ScheduledPromise;
 
 import java.util.ArrayList;
 import java.util.List;
