+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoop.java
@@ -25,8 +25,7 @@ package org.traffichunter.titan.core.concurrent;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
-import org.traffichunter.titan.core.util.concurrent.Promise;
-import org.traffichunter.titan.core.util.concurrent.ScheduledPromise;
+
 import org.traffichunter.titan.core.util.event.EventLoopConstants;
 
 /**
