+++ core/src/main/java/org/traffichunter/titan/core/concurrent/ScheduledPromiseImpl.java
@@ -21,13 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.titan.core.util.concurrent;
+package org.traffichunter.titan.core.concurrent;
 
 import io.netty.util.internal.DefaultPriorityQueue;
 import java.util.concurrent.Callable;
 import java.util.concurrent.Delayed;
 import java.util.concurrent.TimeUnit;
-import org.traffichunter.titan.core.concurrent.EventLoop;
+
 import org.traffichunter.titan.core.util.Time;
 
 /**
